requirements.txt
Flask==2.3.2
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.1
seaborn==0.12.2
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
statsmodels
sympy==1.12
scipy==1.11.3
gunicorn==20.1.0
python-dotenv==1.0.0